chat-service/requirements.txt
fastapi-0.115.6
pydantic-2.10.4
sqlalchemy-2.0.36
aiohttp-3.11.11
asyncio-3.4.3
websocket-0.2.1
python-decouple-3.8
python-dotenv-1.0.1